+++ packages/widgetbook/lib/src/args/double_arg.dart
@@ -14,16 +14,6 @@ class DoubleArg extends Arg<double> with SingleFieldOnly {
       initialValue: value,
     );
   }
-
-  @override
-  DoubleArg init({
-    required String name,
-  }) {
-    return DoubleArg(
-      value,
-      name: $name ?? name,
-    );
-  }
 }
 
 class NullableDoubleArg extends Arg<double?> with SingleFieldOnly {
@@ -37,16 +27,6 @@ class NullableDoubleArg extends Arg<double?> with SingleFieldOnly {
     return DoubleInputField(
       name: 'value',
       initialValue: value ?? 0.0,
-    );
-  }
-
-  @override
-  NullableDoubleArg init({
-    required String name,
-  }) {
-    return NullableDoubleArg(
-      value,
-      name: $name ?? name,
     );
   }
 }